+++ lib/change_health/models/claim/submission/provider.rb
@@ -3,9 +3,11 @@ module ChangeHealth
     module Claim
       class Provider < Hashie::Trash
         property :address, required: false
+        property :contactInformation, from: :contact_information, required: false
         property :employerId, from: :employer_id, required: false # or ssn
         property :firstName, from: :first_name, required: false
         property :lastName, from: :last_name, required: false
+        property :organizationName, from: :organization_name, required: false
         property :npi, required: false
         property :providerType, from: :provider_type, required: false
         property :ssn, required: false # or employer id