+++ test/change_health/request/submission_test.rb
@@ -2,7 +2,7 @@ require 'test_helper'
 
 class SubmissionTest < Minitest::Test
   describe 'claim_submission' do
-    let(:claim_submission) { ChangeHealth::Models::Claim::Submission.new }
+    let(:claim_submission) { ChangeHealth::Request::Claim::Submission.new }
 
     describe 'object' do
       describe 'serializes' do
@@ -31,7 +31,7 @@ class SubmissionTest < Minitest::Test
     describe 'api' do
       describe '#health_check' do
         let(:response) { build_response(file: 'health_check.response.json') }
-        let(:health_check_endpoint) { ChangeHealth::Models::Claim::Submission::HEALTH_CHECK_ENDPOINT }
+        let(:health_check_endpoint) { ChangeHealth::Request::Claim::Submission::HEALTH_CHECK_ENDPOINT }
 
         it 'calls health check' do
           stub_change_health(endpoint: health_check_endpoint, response: response, verb: :get)
@@ -44,7 +44,7 @@ class SubmissionTest < Minitest::Test
 
       describe '#submission mock' do
         let(:response) { build_response(file: 'claim/submission/success.example.response.json') }
-        let(:health_check_endpoint) { ChangeHealth::Models::Claim::Submission::SUBMISSION_ENDPOINT }
+        let(:health_check_endpoint) { ChangeHealth::Request::Claim::Submission::SUBMISSION_ENDPOINT }
 
         before do
           stub_change_health(endpoint: health_check_endpoint, response: response)